Fix FormatCast bug
Fix FormatCast bug.

+++ Bia/biaMachineContext.hpp
@@ -55,7 +55,7 @@ BiaMachineContext(std::shared_ptr<BiaAllocator> p_pAllocator) : m_storage(34), m
 
 	m_index.insert({ "lul", pasd });
 
-	auto lul = [](int32_t i) {
+	auto lul = [](long i) {
 		printf("hiasd askdw %i\n", i);
 	};
 	auto wd = m_storage.CreateElement<framework::executable::BiaLambdaFunction<decltype(lul)>>(std::move(lul));
